rework rcmultispg
The DAQ Threads are now device readers which package up their measurements
and drop them into a Queue.  A separate writer Thread reads this Queue and
is responsible for disk I/O. The main function supervises these Threads and
coordinates startup/shutdown.

This new structure makes it easy to save raw data to dedicated files in real
time if so desired, and to catch SIGUSR1 to trigger an immediate dump of the
spectrogram.