checkpoint a whole lot of work on rcmultispg
It's not so much a spectrogram generator now, as it is a multithreaded data
logger that can geotag multiple data streams. Generation of spectrograms,
tracks, pretty pictures, etc. will be left to some other program.

This of course meant a whole lot of rototilling, moving stuff around to
make things easier to share and test.