switch from namedtuples to dataclasses - they're easier to use.
this now requires python >=3.10 for "kwargs_only" support in dataclasses
which has the advantage of making all the argument names explicit, and
allowing the members to be type-checked.

Since I had to change all of the calls to create thes custom containers,
I also took the opportunity to enforce type checks, and renamed all the
"datetime" members to not shadow datetime.datetime.

Then I had to add a custom json encoder to support my weird dataclasses,
and seemingly simple things like datetimes...