I really do need to start hardware-in-the-loop testing
Discovered by running rcmultispg and finding out that the Spectrum from the
radiacode library was different from rc_types.Spectrum with the encoder
mixin the hard way.

This motivated me to make the JSON encoder use `as_dict` if an object has
such a method, and implement the same encoder for dataclasses without one.

And I finally got around to cleaning up all the "time" fields in the logger,
replacing them with "dt" to transport a datetime as reported by the sender
(GPS, Radiacode, etc), and "monotime" storing time.monotonic() to act as a
correlation identifier because time absolutely will jump around when I'mx
using the logger out in the wild. Many ARM SBCs don't have a proper RTC and
network might not be available to sync time via NTP, both of which mean that
the Radiacode time will almost certainly be incorrect; the GPS receiver may
need a while to start producing valid solutions - as long as all the logs have
the monotonic time counter, a single valid time measurement can be used to
repair all the other timestamps.